# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: VELLA E-commerce API
  version: 0.1.0
  description: B2C single-seller e-commerce platform. Phase 1 covers
    authentication; catalog, cart, checkout, orders and admin are added in later
    slices.
servers:
  - url: http://localhost:4000
    description: Local dev
tags:
  - name: Auth
    description: Registration + session management
  - name: Catalog
    description: Public categories and products
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              pattern: ^[A-Z][A-Z0-9_]*$
            message:
              type: string
            details:
              nullable: true
          required:
            - code
            - message
      required:
        - error
    UserDto:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum:
            - CUSTOMER
            - ADMIN
            - SUPER_ADMIN
      required:
        - id
        - email
        - role
    CartMergeResult:
      type: object
      properties:
        truncatedItems:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: string
              requested:
                type: integer
                minimum: 0
                exclusiveMinimum: true
              granted:
                type: integer
                minimum: 0
            required:
              - skuId
              - requested
              - granted
          description: SKUs whose summed qty was clamped down to current stock
        droppedItems:
          type: array
          items:
            type: object
            properties:
              skuId:
                type: string
              reason:
                type: string
                enum:
                  - INACTIVE
                  - OUT_OF_STOCK
            required:
              - skuId
              - reason
          description: SKUs from the guest cart that were silently dropped during merge
      required:
        - truncatedItems
        - droppedItems
    AuthSuccess:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserDto"
        accessToken:
          type: string
          description: JWT access token, 15 min TTL — store in memory only
        cartMergeResult:
          allOf:
            - $ref: "#/components/schemas/CartMergeResult"
            - description: Guest-cart merge outcome — empty arrays when no guest cookie or
                nothing to merge
      required:
        - user
        - accessToken
        - cartMergeResult
    RegisterBody:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          format: email
        password:
          type: string
          minLength: 8
          maxLength: 72
      required:
        - email
        - password
    LoginBody:
      type: object
      properties:
        email:
          type: string
          maxLength: 254
          format: email
        password:
          type: string
          minLength: 1
          maxLength: 72
      required:
        - email
        - password
    CategoryDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        parentId:
          type: string
          nullable: true
      required:
        - id
        - name
        - slug
        - parentId
    CategoryListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CategoryDto"
      required:
        - items
    ProductImageDto:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        alt:
          type: string
          nullable: true
        sort:
          type: integer
      required:
        - id
        - url
        - alt
        - sort
    VariantOptionDto:
      type: object
      properties:
        id:
          type: string
        value:
          type: string
        sort:
          type: integer
      required:
        - id
        - value
        - sort
    VariantDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              value:
                type: string
              sort:
                type: integer
            required:
              - id
              - value
              - sort
      required:
        - id
        - name
        - options
    SkuDto:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        price:
          type: string
          pattern: ^-?\d+(\.\d{1,2})?$
        stock:
          type: integer
          minimum: 0
        optionCombination:
          type: object
          additionalProperties:
            type: string
        status:
          type: string
          enum: &a2
            - ACTIVE
            - ARCHIVED
      required:
        - id
        - code
        - price
        - stock
        - optionCombination
        - status
    ProductListItemDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        basePrice:
          type: string
          pattern: ^-?\d+(\.\d{1,2})?$
        status:
          type: string
          enum: &a1
            - DRAFT
            - ACTIVE
            - ARCHIVED
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
          required:
            - id
            - name
            - slug
        primaryImage:
          type: object
          nullable: true
          properties:
            id:
              type: string
            url:
              type: string
              format: uri
            alt:
              type: string
              nullable: true
            sort:
              type: integer
          required:
            - id
            - url
            - alt
            - sort
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - basePrice
        - status
        - category
        - primaryImage
        - createdAt
    ProductDetailDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        basePrice:
          type: string
          pattern: ^-?\d+(\.\d{1,2})?$
        status:
          type: string
          enum: *a1
        category:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            slug:
              type: string
          required:
            - id
            - name
            - slug
        images:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              url:
                type: string
                format: uri
              alt:
                type: string
                nullable: true
              sort:
                type: integer
            required:
              - id
              - url
              - alt
              - sort
        variants:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              options:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    value:
                      type: string
                    sort:
                      type: integer
                  required:
                    - id
                    - value
                    - sort
            required:
              - id
              - name
              - options
        skus:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              code:
                type: string
              price:
                type: string
                pattern: ^-?\d+(\.\d{1,2})?$
              stock:
                type: integer
                minimum: 0
              optionCombination:
                type: object
                additionalProperties:
                  type: string
              status:
                type: string
                enum: *a2
            required:
              - id
              - code
              - price
              - stock
              - optionCombination
              - status
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - description
        - basePrice
        - status
        - category
        - images
        - variants
        - skus
        - createdAt
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              slug:
                type: string
              name:
                type: string
              basePrice:
                type: string
                pattern: ^-?\d+(\.\d{1,2})?$
              status:
                type: string
                enum: *a1
              category:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  slug:
                    type: string
                required:
                  - id
                  - name
                  - slug
              primaryImage:
                type: object
                nullable: true
                properties:
                  id:
                    type: string
                  url:
                    type: string
                    format: uri
                  alt:
                    type: string
                    nullable: true
                  sort:
                    type: integer
                required:
                  - id
                  - url
                  - alt
                  - sort
              createdAt:
                type: string
                format: date-time
            required:
              - id
              - slug
              - name
              - basePrice
              - status
              - category
              - primaryImage
              - createdAt
        nextCursor:
          type: string
          nullable: true
      required:
        - items
        - nextCursor
  parameters: {}
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new customer + auto-login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created + access token issued + refresh cookie set
          headers:
            Set-Cookie:
              description: HttpOnly refresh cookie (vella_refresh)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: EMAIL_TAKEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: RATE_LIMITED (3 registers / day / IP)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Authenticate with email + password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Access token issued + refresh cookie set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: INVALID_CREDENTIALS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: RATE_LIMITED (5 logins / minute / IP)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Rotate the refresh cookie + issue a fresh access token
      description: Reads vella_refresh cookie. On success the response Set-Cookie
        carries a new refresh value and the old one is revoked. Reuse of a stale
        token revokes the entire family.
      responses:
        "200":
          description: Rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSuccess"
        "401":
          description: INVALID_TOKEN / TOKEN_EXPIRED / TOKEN_REUSED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: TOKEN_RACED (concurrent rotation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/categories:
    get:
      tags:
        - Catalog
      summary: List all product categories (flat, sorted by name)
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryListResponse"
  /api/products:
    get:
      tags:
        - Catalog
      summary: List ACTIVE products with cursor pagination
      description: Filterable by category slug. `cursor` echoes the previous
        response's `nextCursor`. A null nextCursor means the last page has been
        returned. A cursor pointing at a since-deleted product yields an empty
        page (not an error).
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 80
            description: Restrict results to the given category slug
            example: footwear
          required: false
          name: categorySlug
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
            description: Page size (1–50)
            example: 12
          required: false
          name: limit
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 64
            description: Opaque cursor from the previous response
          required: false
          name: cursor
          in: query
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/products/{slug}:
    get:
      tags:
        - Catalog
      summary: Get full product detail by slug
      description: Includes all images, variants, options, and ACTIVE SKUs. DRAFT or
        ARCHIVED products return 404 even when the slug matches a real row.
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 80
            example: classic-crew-tee
          required: true
          name: slug
          in: path
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetailDto"
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: PRODUCT_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Revoke the current refresh token + clear cookie
      description: Idempotent. Returns 204 even when no refresh cookie was sent.
      responses:
        "204":
          description: Logged out (or already logged out)
        "429":
          description: RATE_LIMITED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"